Fix callscripts

+++ packages/aragon-wrapper/src/evmscript/index.js
@@ -10,7 +10,7 @@ export const CALLSCRIPT_ID = '0x00000001'
 export function encodeCallScript (actions: Array<CallScriptAction>): string {
   return actions.reduce((script, { to, data }) => {
     const address = abi.encodeParameter('address', to)
-    const dataLength = abi.encodeParameter('uint32', (data.length - 2) / 2).toString('hex')
+    const dataLength = abi.encodeParameter('uint256', (data.length - 2) / 2).toString('hex')
 
     return script + address.slice(26) + dataLength.slice(58) + data.slice(2)
   }, CALLSCRIPT_ID)
